+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Support for register data transaction.
+
 ## 0.5.1 2021-11-19
 
 ### Added